Fix S3 stream reading past end of buffer.
The S3 stream reads blocks of data in at a time and hands them out for
individual read requests.  The code was incorrectly computing the number
of bytes remaining in the current block and was reading past the end of
the block instead of triggering the fetch of the next block.